added a few methods for repositories to test Spring Data JPA queries + small fixes with datatype

+++ src/main/java/ua/roman/flats/repositories/name/NameCityRepository.java
@@ -3,5 +3,5 @@ package ua.roman.flats.repositories.name;
 import org.springframework.data.repository.CrudRepository;
 import ua.roman.flats.domain.name.NameCity;
 
-public interface NameCityRepository extends CrudRepository<NameCity, Long> {
+public interface NameCityRepository extends CrudRepository<NameCity, Integer> {
 }
